requirements.txt
torch
torchvision
torchaudio

timm
wandb
ftfy
regex
easydict
pyyaml
textaugment
ipdb

torchmetrics
matplotlib
jupyter
ipykernel
